Some refactoring and improvements
