+++ backend/src/main/kotlin/se/magello/plugins/Routing.kt
@@ -18,6 +18,9 @@ import io.ktor.server.routing.routing
 import kotlinx.serialization.Serializable
 import mu.KotlinLogging
 import se.magello.cinode.CinodeClient
+import se.magello.db.repositories.SkillRepository
+import se.magello.db.repositories.UserRepository
+import se.magello.db.repositories.WorkAssignmentRepository
 import se.magello.map.EniroAddressLookupClient
 import se.magello.salesforce.SalesForceClient
 import se.magello.workflow.UserDataFetcher
@@ -47,18 +50,6 @@ data class Users(val limit: Int = 10, val offset: Int = 0) {
 }
 
 @Serializable
-@Resource("/salesforce")
-class SalesForce {
-    @Serializable
-    @Resource("/versions")
-    data class Versions(val parent: SalesForce = SalesForce(), val limit: Int = 100, val offset: Int = 0)
-
-    @Serializable
-    @Resource("/users")
-    data class Users(val parent: SalesForce = SalesForce())
-}
-
-@Serializable
 @Resource("/skill")
 class Skill {
     @Serializable
@@ -71,17 +62,26 @@ class Skill {
 }
 
 private val logger = KotlinLogging.logger {}
+
+// TODO: clients out and add them to workflow setup
+//   - Create wrapper object for responses: {data: T, errors: []Errors}
+//   - Handle JobIsRunning exceptions in all repositories.
+//   - Create an internal and external model
 fun Application.configureRouting(config: Config) {
     val cinodeConfig = config.getConfig("cinode")
     val salesForceConfig = config.getConfig("salesforce")
     val cinodeClient = CinodeClient(cinodeConfig)
     val salesForceClient = SalesForceClient(salesForceConfig)
-    val client = MergeUserDataWorkflow(UserDataFetcher(cinodeClient, salesForceClient, EniroAddressLookupClient()))
+    val workflow = MergeUserDataWorkflow(UserDataFetcher(cinodeClient, salesForceClient, EniroAddressLookupClient()))
+
+    val userRepository = UserRepository(workflow)
+    val skillRepository = SkillRepository()
+    val workAssignmentRepository = WorkAssignmentRepository(workflow)
 
     val routes = routing {
         get<Workplaces.Workplace> {
             try {
-                val message = client.getWorkAssignment(it.organisationId)
+                val message = workAssignmentRepository.getWorkAssignment(it.organisationId)
                 if (message == null) {
                     call.respond(HttpStatusCode.NotFound)
                 } else {
@@ -93,29 +93,30 @@ fun Application.configureRouting(config: Config) {
         }
         get<Workplaces> {
             try {
-                val message = client.getAllWorkAssignments(it.limit, it.offset)
+                val message = workAssignmentRepository.getAllWorkAssignments(it.limit, it.offset)
                 call.respond(message)
             } catch (je: JobRunningException) {
-                call.respond(HttpStatusCode.ServiceUnavailable)
-            }
-        }
-        get<Users> {
-            try {
-                val message = client.getAllUsers(it.limit, it.offset)
-                call.respond(message)
-            } catch (je: JobRunningException) {
-                // TODO: Make it so we don't have to copy this
                 call.respond(HttpStatusCode.ServiceUnavailable)
             }
         }
         authenticate("azure-jwt") {
+            get<Users> {
+                try {
+                    val allUsers = userRepository.getAllUsers(it.limit, it.offset)
+                    call.respond(allUsers)
+                } catch (je: JobRunningException) {
+                    // TODO: Make it so we don't have to copy this
+                    call.respond(HttpStatusCode.ServiceUnavailable)
+                }
+            }
             get<Users.Self> {
                 when (val principal = call.principal<JWTPrincipal>()) {
                     null -> call.respond(HttpStatusCode.Unauthorized)
+
                     else -> {
-                        when (val message = client.getUserSelf(principal)) {
-                            null -> call.respond(HttpStatusCode.NotFound) // TODO: Return a better response explaining why we cant find the user
-                            else -> call.respond(message)
+                        when (val userSelf = userRepository.getUserSelf(principal)) {
+                            null -> call.respond(HttpStatusCode.NotFound)
+                            else -> call.respond(userSelf)
                         }
                     }
                 }
@@ -126,7 +127,7 @@ fun Application.configureRouting(config: Config) {
                     else -> {
                         val preferences = call.receive<MagelloUserPreferences>()
                         try {
-                            client.postUserPreferences(principal, preferences)
+                            userRepository.postUserPreferences(principal, preferences)
                             call.respond(HttpStatusCode.NoContent)
                         } catch (ie: IllegalStateException) {
                             logger.info { ie.message }
@@ -136,16 +137,10 @@ fun Application.configureRouting(config: Config) {
                 }
             }
             get<Skill.Search> {
-                call.respond(client.searchSkill(it.query))
+                call.respond(skillRepository.searchSkill(it.query))
             }
             get<Skill.Id> {
-                call.respond(client.getUserSkillsForSkillId(it.id))
-            }
-            get<SalesForce.Versions> {
-                call.respond(salesForceClient.getVersions())
-            }
-            get<SalesForce.Users> {
-                call.respond(salesForceClient.getAllMagelloUsers() ?: Any())
+                call.respond(skillRepository.getUserSkillsForSkillId(it.id))
             }
         }
     }
